cleaned up rmf_site_picking
