+++ crates/rmf_site_camera/src/plugins.rs
@@ -12,7 +12,7 @@ use crate::{
 
 /// Plugin to add a blocker to a registry of blockers.
 ///
-/// E.G: [`UiHovered`] -> CameraControlBlockers to block camera controls on ui hovered.
+/// E.G: [`UiFocused`] -> CameraControlBlockers to block camera controls on ui hovered.
 #[derive(Default)]
 pub struct BlockerRegistration<Blocker, Registry>
 where
