+++ crates/rmf_site_picking/Cargo.toml
@@ -24,6 +24,7 @@ smallvec = {workspace = true}
 rmf_site_camera = {workspace = true}
 rmf_site_mesh = {workspace = true}
 rmf_site_animate = {workspace = true}
+bevy_egui = {workspace = true}
 
 [dev-dependencies]
 bevy = {workspace = true, features = ["dynamic_linking"]}
